requirements.txt
Flask==2.2.3
joblib==1.2.0
matplotlib==3.7.0
numpy==1.24.2
Pillow==9.4.0
requests==2.28.2
scikit-learn==1.2.1